pyproject.toml
[tool.poetry]
name = "rpi-reactive-gpio"
version = "0.1.0"
description = "Syntax sugar for controlling RPi.GPIO with reactive design."
authors = ["MarkParker5 <markparker.it@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "rpi_reactive_gpio"}]

[tool.poetry.dependencies]
python = "^3.10"
"RPi.GPIO" = "^0.7.1"

[tools.mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
